requirements.txt
python-telegram-bot==21.6
python-docx==1.1.2
aiohttp==3.10.11
openpyxl==3.1.5
